requirements.txt
scipy==1.17.0
scikit-learn==1.8.0
opencv-python==4.13.0.90
trimesh==4.11.1
einops==0.8.2
jsonlines==4.0.0
transformers==4.57.6
accelerate==1.12.0
datasets==4.5.0
trl==0.27.1
hydra-core==1.3.2
tensorboard==2.20.0
wandb==0.24.0
x-transformers==1.26.6
pybind11==3.0.1
open3d==0.19.0
